+++ mkostemps.c
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
 
+#ifndef __NetBSD__
 #include <stdio.h>
 #include <string.h>
 #include <errno.h>
@@ -55,4 +56,5 @@ exit:
 
 #ifdef __GLIBC__
 weak_alias(mkostemps, mkostemps64);
+#endif
 #endif