+++ mkstemp.c
@@ -1,9 +1,10 @@
 /*
- * Copyright 2021-2023 Gaël PORTAY
+ * Copyright 2021-2024 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
 
+#ifndef __NetBSD__
 #include <stdio.h>
 #include <errno.h>
 
@@ -21,4 +22,5 @@ int mkstemp(char *path)
 
 #ifdef __GLIBC__
 weak_alias(mkstemp, mkstemp64);
+#endif
 #endif