+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - Remove the detailled `execve` output command on level 4 and above
+- Remove the functions `mkdtemp()`, `mkostemp()`, `mkostemps()`, `mkstemp()`,
+  `mkstemps()` and `mktemp()` on [NetBSD]
 
 ### Changed